feat: Adding Matrix lib to the main code
